Refactor: Extract VideoList, QualityRating, and atomic editor components. Implement useProjectState hook and project-service.
