+++ app/page.tsx
@@ -11,7 +11,9 @@ import { Dialog, DialogContent, DialogTitle, DialogHeader, DialogDescription, Di
 import { LightRays } from '@/components/ui/light-rays';
 import { BlurFade } from '@/components/ui/blur-fade';
 import { FinalVideoEditor } from '@/components/FinalVideoEditor';
+import { VideoList } from '@/components/VideoList';
 import { useFinalizeVideo } from '@/hooks/useFinalizeVideo';
+import { useProjectState } from '@/hooks/useProjectState';
 import {
   TransitionVideo,
   FinalVideo,
@@ -19,6 +21,8 @@ import {
   SpeedCurvedBlobCache,
   UpdateReason,
   FinalizeContext,
+  PreflightWarning,
+  VideoMetadata,
 } from '@/lib/types';
 import TextPressure from '@/components/text/text-pressure';
 import { canEncodeVideo, getEncodableVideoCodecs } from 'mediabunny';
@@ -28,6 +32,13 @@ import {
   getPresetBezier,
 } from '@/lib/easing-presets';
 import { DEFAULT_EASING } from '@/lib/speed-curve-config';
+import {
+  readVideoMetadata,
+  getCodecStringForResolution,
+  estimateBitrateForResolution,
+  formatResolutionLabel,
+  runPreflightChecks,
+} from '@/lib/project-service';
 import { AVC_LEVEL_4_0, AVC_LEVEL_5_1 } from '@/lib/video-encoding';
 
 type AudioFinalizeOptions = {
@@ -35,80 +46,6 @@ type AudioFinalizeOptions = {
   audioSettings?: AudioProcessingOptions;
   updateHint?: UpdateReason;
 };
-
-type VideoMetadata = {
-  width: number;
-  height: number;
-  duration: number;
-};
-
-const FOUR_K_WIDTH = 3840;
-const FOUR_K_HEIGHT = 2160;
-const MAX_TOTAL_SIZE_BYTES = 1.5 * 1024 * 1024 * 1024; // 1.5GB
-
-interface PreflightWarning {
-  id: string;
-  title: string;
-  description: string;
-  severity: 'warning' | 'error';
-}
-
-const readVideoMetadata = (file: File | Blob): Promise<VideoMetadata> =>
-  new Promise((resolve, reject) => {
-    const url = URL.createObjectURL(file);
-    const video = document.createElement('video');
-    video.preload = 'metadata';
-    video.muted = true;
-    video.playsInline = true;
-
-    const cleanup = () => {
-      video.removeAttribute('src');
-      video.load();
-      URL.revokeObjectURL(url);
-    };
-
-    video.onloadedmetadata = () => {
-      const width = video.videoWidth;
-      const height = video.videoHeight;
-      const duration = Number.isFinite(video.duration) ? video.duration : 0;
-      cleanup();
-      if (!width || !height) {
-        reject(new Error('Unable to determine video dimensions.'));
-        return;
-      }
-      resolve({ width, height, duration });
-    };
-
-    video.onerror = () => {
-      cleanup();
-      reject(new Error('Failed to read video metadata.'));
-    };
-
-    video.src = url;
-  });
-
-const getCodecStringForResolution = (width: number, height: number) =>
-  width >= FOUR_K_WIDTH || height >= FOUR_K_HEIGHT ? AVC_LEVEL_5_1 : AVC_LEVEL_4_0;
-
-const estimateBitrateForResolution = (width: number, height: number) => {
-  const pixels = width * height;
-  if (pixels >= FOUR_K_WIDTH * FOUR_K_HEIGHT) {
-    return 25_000_000;
-  }
-  if (pixels >= 2560 * 1440) {
-    return 16_000_000;
-  }
-  if (pixels >= 1920 * 1080) {
-    return 12_000_000;
-  }
-  if (pixels >= 1280 * 720) {
-    return 6_000_000;
-  }
-  return 3_000_000;
-};
-
-const formatResolutionLabel = (width?: number, height?: number) =>
-  width && height ? `${width}x${height}` : 'this resolution';
 
 const cloneSegmentForLoop = (
   segment: TransitionVideo,
@@ -189,34 +126,41 @@ const syncSegmentsToLoopCount = (
 };
 
 export default function Home() {
-  const [uploadedVideos, setUploadedVideos] = useState<File[]>([]);
-  const [transitionVideos, setTransitionVideos] = useState<TransitionVideo[]>([]);
-  const [selectedSegmentId, setSelectedSegmentId] = useState<number | null>(null);
-  const [loopCount, setLoopCount] = useState(1);
-  const [draggingVideoIndex, setDraggingVideoIndex] = useState<number | null>(null);
-  const [dropIndicatorIndex, setDropIndicatorIndex] = useState<number | null>(null);
-  const [finalVideo, setFinalVideo] = useState<FinalVideo | null>(null);
-  const [isFinalizingVideo, setIsFinalizingVideo] = useState(false);
-  const [finalizationProgress, setFinalizationProgress] = useState(0);
-  const [finalizationMessage, setFinalizationMessage] = useState('');
+  const {
+    uploadedVideos,
+    transitionVideos,
+    setTransitionVideos,
+    selectedSegmentId,
+    setSelectedSegmentId,
+    loopCount,
+    setLoopCount,
+    finalVideo,
+    setFinalVideo,
+    isFinalizingVideo,
+    setIsFinalizingVideo,
+    finalizationProgress,
+    setFinalizationProgress,
+    finalizationMessage,
+    setFinalizationMessage,
+    speedCurveCache,
+    setSpeedCurveCache,
+    prevAudioBlobRef,
+    prevAudioSettingsRef,
+    addVideos,
+    resetProject,
+    removeVideo,
+    updateSegmentMetadata,
+    cleanupSegmentResources,
+  } = useProjectState();
+
   const [isDropZoneHovered, setIsDropZoneHovered] = useState(false);
   const [isSupported, setIsSupported] = useState(true);
   const [preflightWarnings, setPreflightWarnings] = useState<PreflightWarning[]>([]);
   const [showPreflightDialog, setShowPreflightDialog] = useState(false);
-  const transitionVideosRef = useRef<TransitionVideo[]>([]);
-
-  // Cache for speed-curved blobs to enable fast audio-only updates
-  const [speedCurveCache, setSpeedCurveCache] = useState<SpeedCurvedBlobCache | null>(null);
-
-  // Refs for tracking previous values to detect what changed
-  const prevAudioBlobRef = useRef<Blob | null>(null);
-  const prevAudioSettingsRef = useRef<AudioProcessingOptions | null>(null);
 
   useEffect(() => {
     const checkSupport = async () => {
       try {
-        // Check if the browser supports any video encoding
-        // This implicitly checks for WebCodecs support
         const codecs = await getEncodableVideoCodecs();
         if (codecs.length === 0) {
           setIsSupported(false);
@@ -231,203 +175,20 @@ export default function Home() {
 
   const { finalizeVideos } = useFinalizeVideo();
 
-  const cleanupSegmentResources = useCallback((segments: TransitionVideo[]) => {
-    segments.forEach((segment) => {
-      if (segment.url) {
-        try {
-          URL.revokeObjectURL(segment.url);
-        } catch {
-          // Ignore double-revoke errors
-        }
-      }
-    });
-  }, []);
-
-  useEffect(() => {
-    transitionVideosRef.current = transitionVideos;
-  }, [transitionVideos]);
-
-  useEffect(() => {
-    return () => {
-      cleanupSegmentResources(transitionVideosRef.current);
-    };
-  }, [cleanupSegmentResources]);
-
-  const evaluateVideoEncodeCapability = useCallback(
-    async (segments: TransitionVideo[]) => {
-      await Promise.all(
-        segments.map(async (segment) => {
-          const blobSource = segment.cachedBlob ?? segment.file;
-          if (!blobSource) {
-            return;
-          }
-          const segmentId = segment.id;
-
-          setTransitionVideos((prev) => {
-            if (!prev.some((v) => v.id === segmentId)) {
-              return prev;
-            }
-            return prev.map((v) =>
-              v.id === segmentId
-                ? {
-                  ...v,
-                  encodeCapability: {
-                    status: 'checking',
-                    message: 'Checking device encoder support…',
-                  },
-                }
-                : v
-            );
-          });
-
-          try {
-            const metadata = await readVideoMetadata(blobSource);
-            const codecString = getCodecStringForResolution(metadata.width, metadata.height);
-            const bitrate = estimateBitrateForResolution(metadata.width, metadata.height);
-            const supported = await canEncodeVideo('avc', {
-              width: metadata.width,
-              height: metadata.height,
-              bitrate,
-              fullCodecString: codecString,
-            });
-
-            setTransitionVideos((prev) => {
-              if (!prev.some((v) => v.id === segmentId)) {
-                return prev;
-              }
-              return prev.map((v) =>
-                v.id === segmentId
-                  ? {
-                    ...v,
-                    width: metadata.width,
-                    height: metadata.height,
-                    encodeCapability: {
-                      status: supported ? 'supported' : 'unsupported',
-                      message: supported
-                        ? `Device can encode ${metadata.width}x${metadata.height} AVC`
-                        : `Device encoder cannot output ${metadata.width}x${metadata.height}`,
-                      codecString,
-                      bitrate,
-                    },
-                  }
-                  : v
-              );
-            });
-          } catch (error) {
-            const errorMessage =
-              error instanceof Error ? error.message : 'Unable to verify encode capability';
-            setTransitionVideos((prev) => {
-              if (!prev.some((v) => v.id === segmentId)) {
-                return prev;
-              }
-              return prev.map((v) =>
-                v.id === segmentId
-                  ? {
-                    ...v,
-                    encodeCapability: {
-                      status: 'error',
-                      message: errorMessage,
-                    },
-                  }
-                  : v
-              );
-            });
-          }
-        })
-      );
-    },
-    [setTransitionVideos]
-  );
-
-  const processNewVideoFiles = useCallback(
-    async (files: FileList | null, append: boolean = false) => {
-      if (!(files && files[0])) {
-        return;
-      }
-
-      const videoFiles = Array.from(files).filter((f) => f.type.startsWith('video/'));
-      if (videoFiles.length === 0) return;
-
-      if (append) {
-        setUploadedVideos((prev) => [...prev, ...videoFiles]);
-      } else {
-        setUploadedVideos(videoFiles);
-      }
-
-      try {
-        const nextIdStart = append
-          ? Math.max(0, ...transitionVideos.map((v) => v.id)) + 1
-          : 1;
-
-        const preparedSegments = await Promise.all(
-          videoFiles.map(async (file, index) => {
-            const buffer = await file.arrayBuffer();
-            const cachedBlob = new Blob([buffer], { type: file.type || 'video/mp4' });
-            const objectUrl = URL.createObjectURL(cachedBlob);
-
-            return {
-              id: nextIdStart + index,
-              name: file.name,
-              url: objectUrl,
-              loading: false,
-              duration: 1.5,
-              easingPreset: DEFAULT_EASING,
-              useCustomEasing: false,
-              customBezier: getPresetBezier(DEFAULT_EASING),
-              loopIteration: 1,
-              file,
-              cachedBlob,
-              encodeCapability: {
-                status: 'checking',
-                message: 'Checking device encoder support...',
-              },
-            } as TransitionVideo;
-          })
-        );
-
-        setTransitionVideos((prev) => {
-          if (!append) {
-            cleanupSegmentResources(prev);
-            return preparedSegments;
-          }
-          return [...prev, ...preparedSegments];
-        });
-
-        if (!append) {
-          setLoopCount(1);
-          setSelectedSegmentId(preparedSegments[0]?.id ?? null);
-        } else {
-          // Keep existing selection or select the first new one if nothing selected?
-          // Usually keeping selection is better.
-        }
-
-        // Clear cache as the composition has changed
-        setSpeedCurveCache(null);
-        prevAudioBlobRef.current = null;
-        prevAudioSettingsRef.current = null;
-
-        void evaluateVideoEncodeCapability(preparedSegments);
-      } catch (error) {
-        console.error('Failed to process uploaded videos', error);
-      }
-    },
-    [transitionVideos, cleanupSegmentResources, evaluateVideoEncodeCapability]
-  );
-
   const handleVideosUpload = useCallback(
     (e: React.ChangeEvent<HTMLInputElement>) => {
-      void processNewVideoFiles(e.target.files, false);
+      void addVideos(e.target.files, false);
       if (e.target) e.target.value = '';
     },
-    [processNewVideoFiles]
+    [addVideos]
   );
 
   const handleAddMoreVideos = useCallback(
     (e: React.ChangeEvent<HTMLInputElement>) => {
-      void processNewVideoFiles(e.target.files, true);
+      void addVideos(e.target.files, true);
       if (e.target) e.target.value = '';
     },
-    [processNewVideoFiles]
+    [addVideos]
   );
 
   const handleDragOver = useCallback((e: React.DragEvent) => {
@@ -452,10 +213,10 @@ export default function Home() {
 
       const files = e.dataTransfer.files;
       if (files && files.length > 0) {
-        void processNewVideoFiles(files, false);
+        void addVideos(files, false);
       }
     },
-    [isSupported, processNewVideoFiles]
+    [isSupported, addVideos]
   );
 
   const handleSelectSegment = (id: number) => {
@@ -484,33 +245,6 @@ export default function Home() {
       return nextSegments;
     });
     setLoopCount(nextLoop);
-  };
-
-  const updateSegmentMetadata = (
-    id: number,
-    updates: Partial<TransitionVideo>,
-    applyAll: boolean = false
-  ) => {
-    setTransitionVideos((prev) =>
-      prev.map((segment) => {
-        const shouldUpdate = applyAll || segment.id === id;
-        if (!shouldUpdate) {
-          return segment;
-        }
-
-        const nextUpdates = { ...updates };
-        if (updates.customBezier) {
-          nextUpdates.customBezier = [...updates.customBezier] as [
-            number,
-            number,
-            number,
-            number
-          ];
-        }
-
-        return { ...segment, ...nextUpdates };
-      })
-    );
   };
 
   const handleSegmentDurationChange = (id: number, duration: number, applyAll = false) => {
@@ -542,7 +276,7 @@ export default function Home() {
     if (finalVideo && transitionVideos.length > 0 && selectedSegmentId === null) {
       setSelectedSegmentId(transitionVideos[0].id);
     }
-  }, [finalVideo, transitionVideos, selectedSegmentId]);
+  }, [finalVideo, transitionVideos, selectedSegmentId, setSelectedSegmentId]);
 
   const handleCloneSegmentSettings = (id: number) => {
     const sourceSegment = transitionVideos.find((segment) => segment.id === id);
@@ -607,88 +341,7 @@ export default function Home() {
     await handleFinalizeVideo(undefined, options, true, reason);
   };
 
-  const runPreflightChecks = (segments: TransitionVideo[]): PreflightWarning[] => {
-    const warnings: PreflightWarning[] = [];
 
-    // 1. Check for mixed orientation
-    let hasPortrait = false;
-    let hasLandscape = false;
-    segments.forEach((s) => {
-      if (s.width && s.height) {
-        if (s.height > s.width) hasPortrait = true;
-        else hasLandscape = true;
-      }
-    });
-
-    if (hasPortrait && hasLandscape) {
-      warnings.push({
-        id: 'orientation-mismatch',
-        title: 'Mixed Video Orientations',
-        description: 'You have both portrait and landscape videos. The final output might look rotated or stretched.',
-        severity: 'warning',
-      });
-    }
-
-    // 2. Check for large resolution disparity
-    let minHeight = Infinity;
-    let maxHeight = 0;
-    segments.forEach((s) => {
-      if (s.height) {
-        minHeight = Math.min(minHeight, s.height);
-        maxHeight = Math.max(maxHeight, s.height);
-      }
-    });
-
-    if (maxHeight > 0 && minHeight < Infinity && maxHeight > minHeight * 2) {
-      warnings.push({
-        id: 'resolution-disparity',
-        title: 'Resolution Disparity',
-        description: `Some videos are much larger than others (Max: ${maxHeight}p, Min: ${minHeight}p). Smaller videos will be upscaled, which may look blurry.`,
-        severity: 'warning',
-      });
-    }
-
-    // 3. Check for mixed aspect ratios
-    const validVideos = segments
-      .filter((s) => s.width && s.height)
-      .map((s) => ({ width: s.width!, height: s.height! }));
-
-    if (validVideos.length > 1) {
-      const consistency = calculateAspectRatioConsistency(validVideos);
-      if (consistency < 100) {
-        warnings.push({
-          id: 'aspect-ratio-mismatch',
-          title: 'Mixed Aspect Ratios',
-          description: "Look, you've actually put the wrong file types in. But don't worry about it, we're going to encode it for you so you get a rough idea of what this looks like. But it might give an idea to go back and fix that so that it's consistent.",
-          severity: 'warning',
-        });
-      }
-    }
-
-    // 4. Check total file size
-    const totalSize = segments.reduce((acc, s) => acc + (s.file?.size || 0), 0);
-    if (totalSize > MAX_TOTAL_SIZE_BYTES) {
-      warnings.push({
-        id: 'large-files',
-        title: 'Large Project Size',
-        description: `Total video size is ${(totalSize / (1024 * 1024 * 1024)).toFixed(1)}GB. This might crash the browser during processing.`,
-        severity: 'error', // High risk
-      });
-    }
-
-    // 4. Encoder support (re-using existing status)
-    const unsupportedVideos = segments.filter(s => s.encodeCapability?.status === 'unsupported');
-    if (unsupportedVideos.length > 0) {
-      warnings.push({
-        id: 'unsupported-encoder',
-        title: 'Unsupported Resolution',
-        description: `Your device cannot encode some video resolutions (e.g. ${unsupportedVideos[0].width}x${unsupportedVideos[0].height}). The process will likely fail.`,
-        severity: 'error',
-      });
-    }
-
-    return warnings;
-  };
 
   const handleFinalizeVideo = async (
     segmentsOverride?: TransitionVideo[],
@@ -788,40 +441,7 @@ export default function Home() {
     });
   };
 
-  const handleVideoDragStart = (index: number) => (event: React.DragEvent<HTMLButtonElement>) => {
-    setDraggingVideoIndex(index);
-    event.dataTransfer.effectAllowed = 'move';
-    event.dataTransfer.setData('text/plain', index.toString());
-  };
-
-  const handleVideoDragOver = (index: number) => (event: React.DragEvent<HTMLDivElement>) => {
-    event.preventDefault();
-    if (draggingVideoIndex === null || draggingVideoIndex === index) {
-      setDropIndicatorIndex(null);
-      return;
-    }
-    event.dataTransfer.dropEffect = 'move';
-    setDropIndicatorIndex(index);
-  };
-
-  const handleVideoDrop = (index: number) => (event: React.DragEvent<HTMLDivElement>) => {
-    event.preventDefault();
-    const sourceIndex =
-      draggingVideoIndex ?? Number.parseInt(event.dataTransfer.getData('text/plain'), 10);
-    if (Number.isNaN(sourceIndex)) {
-      setDraggingVideoIndex(null);
-      setDropIndicatorIndex(null);
-      return;
-    }
-    reorderTransitionVideos(sourceIndex, index);
-    setDraggingVideoIndex(null);
-    setDropIndicatorIndex(null);
-  };
-
-  const handleVideoDragEnd = () => {
-    setDraggingVideoIndex(null);
-    setDropIndicatorIndex(null);
-  };
+  // Drag handlers removed (moved to VideoList)
 
   const handlePlayTransitionVideo = (video: TransitionVideo) => {
     if (!video.url || video.loading) {
@@ -995,244 +615,57 @@ export default function Home() {
 
             {/* Uploaded Videos Preview */}
             {uploadedVideos.length > 0 && (
-              <BlurFade delay={0.2} className="w-full">
-                <div className="w-full space-y-6">
-                  <div className="flex items-center justify-between">
-                    <h3 className="text-lg font-semibold">
-                      Uploaded Videos ({uploadedVideos.length})
-                    </h3>
-                    <div className="flex gap-2">
-                      <input
-                        type="file"
-                        accept="video/*"
-                        onChange={(event) => {
-                          void handleAddMoreVideos(event);
-                        }}
-                        className="hidden"
-                        id="add-videos-input"
-                        multiple
-                        disabled={!isSupported}
-                      />
-                      <Button
-                        variant="outline"
-                        size="sm"
-                        onClick={() => document.getElementById('add-videos-input')?.click()}
-                      >
-                        <Plus className="h-4 w-4 mr-2" />
-                        Add Video
-                      </Button>
-                      <Button
-                        variant="outline"
-                        size="sm"
-                        onClick={() => {
-                          setUploadedVideos([]);
-                          setTransitionVideos((prev) => {
-                            cleanupSegmentResources(prev);
-                            return [];
-                          });
-                          setFinalVideo(null);
-                          setSelectedSegmentId(null);
-                          // Clear cache when resetting
-                          setSpeedCurveCache(null);
-                          prevAudioBlobRef.current = null;
-                          prevAudioSettingsRef.current = null;
-                        }}
-                      >
-                        Reset
-                      </Button>
-                    </div>
-                  </div>
-
-                  {/* Videos List with Reordering */}
-                  <div className="space-y-3">
-                    {transitionVideos.map((video, index) => {
-                      const isDragging = draggingVideoIndex === index;
-                      const statusText = video.loading
-                        ? 'Generating...'
-                        : video.error
-                          ? `Error: ${video.error}`
-                          : video.url
-                            ? 'Ready to finalize'
-                            : 'Pending';
-                      const statusColor = video.error
-                        ? 'text-destructive'
-                        : video.loading
-                          ? 'text-muted-foreground'
-                          : 'text-emerald-500';
-                      const encodeCapability = video.encodeCapability;
-                      let encodeStatusText: string | null = null;
-                      let encodeStatusClass = 'text-muted-foreground';
-
-                      if (encodeCapability) {
-                        switch (encodeCapability.status) {
-                          case 'pending':
-                          case 'checking':
-                            encodeStatusText =
-                              encodeCapability.message ?? 'Checking device encoder support…';
-                            encodeStatusClass = 'text-muted-foreground';
-                            break;
-                          case 'supported':
-                            encodeStatusText = null;
-                            break;
-                          case 'unsupported':
-                            encodeStatusText =
-                              encodeCapability.message ??
-                              `Cannot encode ${formatResolutionLabel(video.width, video.height)} on this device`;
-                            encodeStatusClass = 'text-amber-600';
-                            break;
-                          case 'error':
-                            encodeStatusText =
-                              encodeCapability.message ?? 'Encoder support check failed';
-                            encodeStatusClass = 'text-amber-600';
-                            break;
-                          default:
-                            encodeStatusText = encodeCapability.message ?? null;
-                            encodeStatusClass = 'text-muted-foreground';
-                        }
-                      }
-
-                      return (
-                        <div key={video.id}>
-                          <div
-                            className={cn(
-                              'flex items-center gap-3 rounded-lg border border-border/80 bg-secondary/50 p-4 transition-colors',
-                              isDragging && 'ring-2 ring-primary/40 bg-secondary'
-                            )}
-                            onDragOver={handleVideoDragOver(index)}
-                            onDrop={handleVideoDrop(index)}
-                          >
-
-
-                            <button
-                              type="button"
-                              className="flex h-8 w-8 items-center justify-center rounded-md border border-dashed border-border/70 text-muted-foreground transition-colors hover:text-foreground focus-visible:outline focus-visible:outline-2 focus-visible:outline-primary cursor-grab"
-                              draggable
-                              onDragStart={handleVideoDragStart(index)}
-                              onDragEnd={handleVideoDragEnd}
-                              aria-label={`Reorder ${video.name}`}
-                            >
-                              <GripVertical className="h-4 w-4" />
-                            </button>
-
-                            <Button
-                              type="button"
-                              variant="ghost"
-                              size="icon"
-                              className="text-primary hover:bg-primary/10 disabled:text-muted-foreground"
-                              onClick={() => handlePlayTransitionVideo(video)}
-                              disabled={!video.url || video.loading}
-                            >
-                              <PlayCircle className="h-5 w-5" />
-                            </Button>
-
-                            <div className="flex-1 min-w-0 flex items-center gap-3">
-                              {video.url && (
-                                <video
-                                  src={video.url}
-                                  className="h-12 w-16 rounded-md object-cover flex-shrink-0 bg-secondary"
-                                />
-                              )}
-                              <div className="flex-1 min-w-0">
-                                <p className="font-medium text-sm truncate">
-                                  {index + 1}. {video.name}
-                                </p>
-                                <p className={cn('text-xs mt-1', statusColor)}>{statusText}</p>
-                                {encodeStatusText && (
-                                  <p className={cn('text-xs mt-0.5', encodeStatusClass)}>
-                                    {encodeStatusText}
-                                  </p>
-                                )}
-                              </div>
-                            </div>
-
-                            <Button
-                              type="button"
-                              variant="ghost"
-                              size="icon"
-                              className="text-muted-foreground hover:text-foreground hover:bg-secondary/80"
-                              onClick={() => {
-                                setTransitionVideos((prev) => {
-                                  const target = prev.find((v) => v.id === video.id);
-                                  if (target) {
-                                    cleanupSegmentResources([target]);
-                                  }
-                                  return prev.filter((v) => v.id !== video.id);
-                                });
-                                setUploadedVideos((prev) => prev.filter((f) => f.name !== video.name));
-                              }}
-                            >
-                              <Trash2 className="h-4 w-4" />
-                            </Button>
-                          </div>
-                          {dropIndicatorIndex === index && (
-                            <div className="h-0.5 bg-primary mt-2 mb-1" />
-                          )}
-                        </div>
-                      );
-                    })}
-                  </div>
-
-                  {/* Finalize Button for Uploaded Videos */}
-                  {transitionVideos.every((v) => v.url && !v.loading) && !isFinalizingVideo && (
-                    <div className="space-y-3">
-                      <div className="flex justify-center">
-                        <Button
-                          size="lg"
-                          onClick={() => handleFinalizeVideo()}
-                          className="gap-2"
-                        >
-                          <Play className="h-4 w-4" />
-                          {finalVideo ? 'Finalize Again' : 'Finalize & Stitch Videos'}
-                        </Button>
-                      </div>
-                      {hasPendingEncodeChecks && encodeWarnings.length === 0 && (
-                        <p className="text-center text-xs text-muted-foreground">
-                          Checking device encoder support for uploaded videos…
-                        </p>
-                      )}
-                      {encodeWarnings.length > 0 && (
-                        <div className="rounded-lg border border-amber-500/40 bg-amber-500/10 p-4 text-sm text-amber-800 dark:text-amber-200">
-                          <p className="font-semibold">This device can&apos;t encode:</p>
-                          <ul className="mt-2 list-disc space-y-1 pl-5">
-                            {encodeWarnings.map((warning) => (
-                              <li key={warning}>{warning}</li>
-                            ))}
-                          </ul>
-                          <p className="mt-3 text-xs">
-                            Consider downscaling or trimming before finalizing to avoid encoder errors on this device.
-                          </p>
-                        </div>
-                      )}
-                      {encodeErrors.length > 0 && (
-                        <div className="rounded-lg border border-yellow-500/40 bg-yellow-500/10 p-3 text-xs text-yellow-800 dark:text-yellow-200">
-                          Unable to verify encoder support for {encodeErrors.join(', ')}. Finalization may still work,
-                          but it could fail if the device encoder is limited.
-                        </div>
-                      )}
-                    </div>
-                  )}
-
-                  {/* Finalization Progress */}
-                  {isFinalizingVideo && (
-                    <div className="w-full space-y-3">
-                      <div className="flex items-center justify-between">
-                        <p className="text-sm font-semibold text-foreground">
-                          {finalizationMessage}
-                        </p>
-                        <span className="text-sm text-muted-foreground">
-                          {Math.round(finalizationProgress)}%
-                        </span>
-                      </div>
-                      <div className="w-full bg-secondary rounded-full h-2 overflow-hidden">
-                        <div
-                          className="bg-primary h-full transition-all duration-300"
-                          style={{ width: `${finalizationProgress}%` }}
-                        />
-                      </div>
-                    </div>
-                  )}
-                </div>
-              </BlurFade>
+              <VideoList
+                uploadedVideosCount={uploadedVideos.length}
+                transitionVideos={transitionVideos}
+                isSupported={isSupported}
+                isFinalizing={isFinalizingVideo}
+                finalVideo={finalVideo}
+                onAddVideos={handleAddMoreVideos}
+                onReset={() => {
+                  setUploadedVideos([]);
+                  setTransitionVideos((prev) => {
+                    cleanupSegmentResources(prev);
+                    return [];
+                  });
+                  setFinalVideo(null);
+                  setSelectedSegmentId(null);
+                  // Clear cache when resetting
+                  setSpeedCurveCache(null);
+                  prevAudioBlobRef.current = null;
+                  prevAudioSettingsRef.current = null;
+                }}
+                onRemoveVideo={(id, name) => {
+                  setTransitionVideos((prev) => {
+                    const target = prev.find((v) => v.id === id);
+                    if (target) {
+                      cleanupSegmentResources([target]);
+                    }
+                    return prev.filter((v) => v.id !== id);
+                  });
+                  setUploadedVideos((prev) => prev.filter((f) => f.name !== name));
+                }}
+                onPlayVideo={handlePlayTransitionVideo}
+                onReorder={reorderTransitionVideos}
+                onFinalize={() => void handleFinalizeVideo()}
+                encodeWarnings={Array.from(new Set(
+                  transitionVideos
+                    .filter(v => v.encodeCapability?.status === 'unsupported')
+                    .map(v => v.encodeCapability?.message)
+                    .filter(Boolean) as string[]
+                ))}
+                encodeErrors={Array.from(new Set(
+                  transitionVideos
+                    .filter(v => v.encodeCapability?.status === 'error')
+                    .map(v => v.encodeCapability?.message)
+                    .filter(Boolean) as string[]
+                ))}
+                hasPendingEncodeChecks={transitionVideos.some(
+                  v => v.encodeCapability?.status === 'checking' || v.encodeCapability?.status === 'pending'
+                )}
+                finalizationProgress={finalizationProgress}
+                finalizationMessage={finalizationMessage}
+              />
             )}
 
           </>
