+++ components/FinalVideoEditor.tsx
@@ -2,25 +2,10 @@
 
 import { ChangeEvent, memo, useCallback, useEffect, useMemo, useRef, useState } from 'react';
 import { FinalVideo, TransitionVideo, AudioProcessingOptions, UpdateReason } from '@/lib/types';
-import { Label } from '@/components/ui/label';
-import { calculateAspectRatioConsistency } from '@/lib/utils';
-import { CubicBezierEditor } from '@/components/CubicBezierEditor';
 import { Button } from '@/components/ui/button';
-import {
-  Dialog,
-  DialogContent,
-  DialogDescription,
-  DialogFooter,
-  DialogHeader,
-  DialogTitle,
-} from '@/components/ui/dialog';
-import {
-  Tooltip,
-  TooltipContent,
-  TooltipProvider,
-  TooltipTrigger,
-} from '@/components/ui/tooltip';
-import { Loader2, Play } from 'lucide-react';
+import { Play } from 'lucide-react';
+import { getPresetBezier } from '@/lib/easing-presets';
+import { useVideoPlayback } from '@/hooks/useVideoPlayback';
 import { getPresetBezier } from '@/lib/easing-presets';
 import { useVideoPlayback } from '@/hooks/useVideoPlayback';
 import { VideoPlaybackControls } from '@/components/VideoPlaybackControls';
@@ -34,6 +19,10 @@ import { AudioUploadBox } from '@/components/AudioUploadBox';
 import { AudioWaveformVisualization } from '@/components/AudioWaveformVisualization';
 import { useAudioVisualization } from '@/hooks/useAudioVisualization';
 import { Tabs, TabsContent, TabsList, TabsTrigger } from '@/components/ui/tabs';
+import { QualityRating } from '@/components/QualityRating';
+import { UpdatePromptDialog } from '@/components/UpdatePromptDialog';
+import { AudioSettingsPanel } from '@/components/AudioSettingsPanel';
+import { SegmentSettingsPanel } from '@/components/SegmentSettingsPanel';
 
 const LOOP_OPTIONS = [1, 2, 3] as const;
 const BEZIER_THROTTLE_MS = 75;
@@ -341,178 +330,6 @@ function FinalVideoEditorComponent({
     </div>
   );
 
-  const AudioSettingsContent = (
-    <div className="space-y-6">
-      <div>
-        <h4 className="text-2xl font-bold text-foreground">Audio Settings</h4>
-        <p className="text-xs text-muted-foreground">
-          Shape fade envelopes and looping for the background track.
-        </p>
-      </div>
-
-      <div className="space-y-2">
-        <Label htmlFor="audio-fade-in">Fade in (sec)</Label>
-        <div className="flex items-center gap-3">
-          <input
-            id="audio-fade-in"
-            type="range"
-            min={0}
-            max={10}
-            step={0.1}
-            value={audioSettings.fadeIn}
-            onChange={(event) => updateAudioSetting('fadeIn', Number(event.target.value))}
-            className="h-2 flex-1 cursor-pointer rounded-full bg-primary/30"
-          />
-          <input
-            type="number"
-            min={0}
-            max={10}
-            step={0.1}
-            value={audioSettings.fadeIn}
-            onChange={(event) => updateAudioSetting('fadeIn', Number(event.target.value))}
-            className="w-20 rounded-md border border-border bg-background px-2 py-1 text-sm"
-          />
-        </div>
-      </div>
-
-      <div className="space-y-2">
-        <Label htmlFor="audio-fade-out">Fade out (sec)</Label>
-        <div className="flex items-center gap-3">
-          <input
-            id="audio-fade-out"
-            type="range"
-            min={0}
-            max={10}
-            step={0.1}
-            value={audioSettings.fadeOut}
-            onChange={(event) => updateAudioSetting('fadeOut', Number(event.target.value))}
-            className="h-2 flex-1 cursor-pointer rounded-full bg-primary/30"
-          />
-          <input
-            type="number"
-            min={0}
-            max={10}
-            step={0.1}
-            value={audioSettings.fadeOut}
-            onChange={(event) => updateAudioSetting('fadeOut', Number(event.target.value))}
-            className="w-20 rounded-md border border-border bg-background px-2 py-1 text-sm"
-          />
-        </div>
-      </div>
-
-      <Button
-        size="sm"
-        onClick={handleVideoUpdate}
-        disabled={isUpdating}
-        className="gap-2 w-full"
-      >
-        {isUpdating && <Loader2 className="h-4 w-4 animate-spin" />}
-        {isUpdating ? 'Updating.' : 'Update Video'}
-      </Button>
-    </div>
-  );
-
-  const SegmentSettingsContent = selectedSegment ? (
-    <>
-      <div>
-        <div className="flex items-center gap-3">
-          <h4 className="text-2xl font-bold text-foreground">{selectedSegment.name}</h4>
-          {selectedSegment.loopIteration && selectedSegment.loopIteration > 1 && (
-            <span className="rounded-full border border-border/70 bg-background px-2 py-0.5 text-[11px] font-semibold uppercase tracking-wide text-muted-foreground">
-              Loop {selectedSegment.loopIteration}
-            </span>
-          )}
-        </div>
-        <p className="text-xs text-muted-foreground">Fine-tune duration and easing curve.</p>
-      </div>
-
-      <div className="space-y-2">
-        <Label htmlFor="segment-duration">Duration (sec)</Label>
-        <div className="flex items-center gap-3">
-          <input
-            id="segment-duration"
-            type="range"
-            min={0.5}
-            max={6}
-            step={0.01}
-            value={selectedSegment.duration ?? 1.5}
-            onChange={(event) =>
-              onDurationChange(selectedSegment.id, Number(event.target.value), applyAll)
-            }
-            className="h-2 flex-1 cursor-pointer rounded-full bg-primary/30"
-          />
-          <input
-            type="number"
-            min={0.1}
-            step={0.01}
-            value={(selectedSegment.duration ?? 1.5).toFixed(2)}
-            onChange={(event) =>
-              onDurationChange(selectedSegment.id, Number(event.target.value), applyAll)
-            }
-            className="w-20 rounded-md border border-border bg-background px-2 py-1 text-sm"
-          />
-        </div>
-      </div>
-
-      <div className="space-y-3">
-        <div className="flex items-center justify-between gap-2">
-          <Label>Ease Curve</Label>
-          <select
-            id="preset-select"
-            value={selectedSegment.easingPreset ?? ''}
-            onChange={(event) => onPresetChange(selectedSegment.id, event.target.value, applyAll)}
-            className="rounded-md border border-border bg-background py-2 pl-3 pr-8 text-sm"
-          >
-            {easingOptions.map((preset) => (
-              <option key={preset} value={preset}>
-                {preset}
-              </option>
-            ))}
-          </select>
-        </div>
-        <CubicBezierEditor
-          value={curveValue}
-          onChange={handleBezierChange}
-          onCommit={handleBezierCommit}
-        />
-        <p className="text-xs text-muted-foreground">
-          Drag the control points to sculpt a bespoke ease-in/ease-out profile for this segment.
-        </p>
-      </div>
-
-      <div className="space-y-2 pt-2">
-        <label className="flex items-center gap-2 text-sm text-foreground/90">
-          <input
-            type="checkbox"
-            checked={applyAll}
-            onChange={(event) => {
-              const nextValue = event.target.checked;
-              setApplyAll(nextValue);
-              if (nextValue && selectedSegment) {
-                onCloneSegmentSettings(selectedSegment.id);
-              }
-            }}
-            className="h-4 w-4 rounded border-border text-primary focus:ring-primary"
-          />
-          Apply settings to all videos
-        </label>
-        <Button
-          size="sm"
-          onClick={handleVideoUpdate}
-          disabled={isUpdating}
-          className="gap-2 w-full mt-2"
-        >
-          {isUpdating && <Loader2 className="h-4 w-4 animate-spin" />}
-          {isUpdating ? 'Updating…' : 'Update Video'}
-        </Button>
-      </div>
-    </>
-  ) : (
-    <div className="flex h-full flex-col items-center justify-center text-center text-sm text-muted-foreground">
-      <p>Select a segment in the timeline to edit its timing and ease curve.</p>
-    </div>
-  );
-
   return (
     <>
       <div className="w-full h-full flex flex-col lg:flex-row gap-2 lg:gap-6 max-w-[1800px] mx-auto">
@@ -539,58 +356,7 @@ function FinalVideoEditorComponent({
               createdAt={finalVideo.createdAt}
               actions={
                 <div className="flex flex-wrap items-center gap-2 md:gap-4 text-xs font-semibold uppercase tracking-wide text-muted-foreground">
-                  {(() => {
-                    const validSegments = segments
-                      .filter((s) => s.width && s.height)
-                      .map((s) => {
-                        const r = s.width! / s.height!;
-                        return { r: Number(r.toFixed(2)), label: `${s.width}x${s.height}` };
-                      });
-
-                    if (validSegments.length <= 1) return null;
-
-                    const counts: Record<number, number> = {};
-                    for (const { r } of validSegments) {
-                      counts[r] = (counts[r] || 0) + 1;
-                    }
-
-                    const entries = Object.entries(counts).sort((a, b) => b[1] - a[1]);
-                    const maxCount = entries[0][1];
-                    const consistency = Math.round((maxCount / validSegments.length) * 100);
-
-                    if (consistency < 100) {
-                      // Generate breakdown string
-                      const breakdown = entries.map(([ratio, count]) => {
-                        const rVal = Number(ratio);
-                        let label = `${rVal}:1`;
-                        if (Math.abs(rVal - 1.78) < 0.05) label = "16:9 (Landscape)";
-                        else if (Math.abs(rVal - 0.56) < 0.05) label = "9:16 (Portrait)";
-                        else if (Math.abs(rVal - 1.0) < 0.05) label = "1:1 (Square)";
-                        else if (Math.abs(rVal - 1.33) < 0.05) label = "4:3";
-
-                        return `${count} video${count === 1 ? '' : 's'} are ${label}`;
-                      }).join(', ');
-
-                      return (
-                        <TooltipProvider>
-                          <Tooltip>
-                            <TooltipTrigger asChild>
-                              <div
-                                className="flex items-center gap-1.5 rounded-full bg-yellow-500/10 px-2.5 py-1 text-yellow-600 border border-yellow-500/20 cursor-help"
-                              >
-                                <span className="h-1.5 w-1.5 rounded-full bg-yellow-500 animate-pulse" />
-                                <span>Quality: {consistency}%</span>
-                              </div>
-                            </TooltipTrigger>
-                            <TooltipContent>
-                              <p className="max-w-xs">Mixed Aspect Ratios detected. Breakdown: {breakdown}. For best results, use consistent aspect ratios.</p>
-                            </TooltipContent>
-                          </Tooltip>
-                        </TooltipProvider>
-                      );
-                    }
-                    return null;
-                  })()}
+                  <QualityRating segments={segments} />
                   <label className="flex items-center gap-1.5 md:gap-2">
                     <span>Loops</span>
                     <select
@@ -684,12 +450,34 @@ function FinalVideoEditorComponent({
 
             <div className="flex-1 overflow-y-auto min-h-[400px] lg:min-h-0">
               <TabsContent value="segments" className="mt-0 h-full space-y-4 data-[state=inactive]:hidden">
-                {SegmentSettingsContent}
+                <SegmentSettingsPanel
+                  selectedSegment={selectedSegment}
+                  easingOptions={easingOptions}
+                  applyAll={applyAll}
+                  onApplyAllChange={(checked) => {
+                    setApplyAll(checked);
+                    if (checked && selectedSegment) {
+                      onCloneSegmentSettings(selectedSegment.id);
+                    }
+                  }}
+                  onDurationChange={onDurationChange}
+                  onPresetChange={onPresetChange}
+                  onBezierChange={handleBezierChange}
+                  onBezierCommit={handleBezierCommit}
+                  curveValue={curveValue}
+                  onUpdateVideo={handleVideoUpdate}
+                  isUpdating={isUpdating}
+                />
               </TabsContent>
 
               <TabsContent value="audio" className="mt-0 h-full space-y-4 data-[state=inactive]:hidden">
                 {waveformData ? (
-                  AudioSettingsContent
+                  <AudioSettingsPanel
+                    audioSettings={audioSettings}
+                    onUpdateSetting={updateAudioSetting}
+                    onCommit={handleVideoUpdate}
+                    isUpdating={isUpdating}
+                  />
                 ) : (
                   <div className="flex h-full flex-col items-center justify-center text-center text-sm text-muted-foreground">
                     <p>Upload an audio track in the timeline to configure audio settings.</p>
@@ -713,23 +501,15 @@ function FinalVideoEditorComponent({
         </aside>
       </div>
 
-      <Dialog open={showUpdatePrompt} onOpenChange={handlePromptOpenChange}>
-        <DialogContent>
-          <DialogHeader>
-            <DialogTitle>{promptCopy.title}</DialogTitle>
-            <DialogDescription>{promptCopy.description}</DialogDescription>
-          </DialogHeader>
-          <DialogFooter>
-            <Button variant="outline" onClick={() => setUpdatePromptReason(null)} disabled={isUpdating}>
-              Later
-            </Button>
-            <Button onClick={handleVideoUpdate} disabled={isUpdating} className="gap-2">
-              {isUpdating && <Loader2 className="h-4 w-4 animate-spin" />}
-              Update video
-            </Button>
-          </DialogFooter>
-        </DialogContent>
-      </Dialog>
+      <UpdatePromptDialog
+        open={showUpdatePrompt}
+        onOpenChange={handlePromptOpenChange}
+        title={promptCopy.title}
+        description={promptCopy.description}
+        onConfirm={handleVideoUpdate}
+        onCancel={() => setUpdatePromptReason(null)}
+        isUpdating={isUpdating}
+      />
     </>
   );
 }
