+++ lib/types.ts
@@ -93,3 +93,16 @@ export interface FinalizeResult {
   finalBlob: Blob;
   speedCurvedCache: SpeedCurvedBlobCache;
 }
+
+export interface VideoMetadata {
+  width: number;
+  height: number;
+  duration: number;
+}
+
+export interface PreflightWarning {
+  id: string;
+  title: string;
+  description: string;
+  severity: 'warning' | 'error';
+}
